同步官方

+++ modules/api/app/model/falcon_portal/host.go
@@ -1,3 +1,17 @@
+// Copyright 2017 Xiaomi, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package falcon_portal
 
 import (
@@ -26,8 +40,8 @@ type Host struct {
 	Ip            string `json:"ip" gorm:"column:ip"`
 	AgentVersion  string `json:"agent_version"  gorm:"column:agent_version"`
 	PluginVersion string `json:"plugin_version"  gorm:"column:plugin_version"`
-	MaintainBegin uint16 `json:"maintain_begin"  gorm:"column:maintain_begin"`
-	MaintainEnd   uint16 `json:"maintain_end"  gorm:"column:maintain_end"`
+	MaintainBegin int64  `json:"maintain_begin"  gorm:"column:maintain_begin"`
+	MaintainEnd   int64  `json:"maintain_end"  gorm:"column:maintain_end"`
 }
 
 func (this Host) TableName() string {
